fix: add more forgiving validation due to price rounding
Before the created order is marked as paid a validation occurs that the
paid price is the same as the price reported by nop for the created order.

These prices *can* differ slighly due to rounding differences (for example
Klarna only allows two decimals). This commit ensures orders are marked as
paid if the prices are "close enough"--currently leq 1% diff.